# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Juridica",
        "version": "1.0 (2024-07-22)"
    },
    "servers": [
        {
            "url": "http://localhost:8081/legal/public/"
        }
    ],
    "paths": {
        "/api/action": {
            "get": {
                "tags": [
                    "Action"
                ],
                "summary": "Get all actions",
                "operationId": "0e230644ef4b723a3bb9a9ceb8bec6b4",
                "parameters": [
                    {
                        "name": "x-token",
                        "in": "header",
                        "description": "Key API",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get account user information"
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                }
            }
        },
        "/api/city": {
            "get": {
                "tags": [
                    "City"
                ],
                "summary": "Get all Cities",
                "operationId": "c2504917c4da1139bffc3dd1d5409f8f",
                "parameters": [
                    {
                        "name": "x-token",
                        "in": "header",
                        "description": "Key API",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get account user information"
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                }
            }
        },
        "/api/class": {
            "get": {
                "tags": [
                    "Class"
                ],
                "summary": "Get all class",
                "operationId": "46403a8934b67b310f227b57e25a3be6",
                "parameters": [
                    {
                        "name": "x-token",
                        "in": "header",
                        "description": "Key API",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get account user information"
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                }
            }
        },
        "/api/office": {
            "get": {
                "tags": [
                    "Office"
                ],
                "summary": "Get all Office",
                "operationId": "9c6bdf78328916c7a52a490ed702d975",
                "parameters": [
                    {
                        "name": "x-token",
                        "in": "header",
                        "description": "Key API",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get account user information"
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                }
            },
            "post": {
                "tags": [
                    "Office"
                ],
                "summary": "Create Office",
                "operationId": "4f56e944a201491f6b9d37b65e0721c8",
                "parameters": [
                    {
                        "name": "x-token",
                        "in": "header",
                        "description": "Key API",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Request Body Description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "examples": {
                                "result": {
                                    "summary": "An result object.",
                                    "value": {
                                        "name": "",
                                        "address": ""
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                }
            }
        },
...
        "/api/common": {
            "get": {
                "tags": [
                    "Common"
                ],
                "summary": "Get actions,offices and failures",
                "operationId": "d5a403ac6bd7b6763a09e3283696fabb",
                "parameters": [
                    {
                        "name": "x-token",
                        "in": "header",
                        "description": "Key API",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get account user information"
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                }
            }
        },
        "/api/person": {
            "get": {
                "tags": [
                    "Person"
                ],
                "summary": "Get all person or Search full_name and Doc",
                "operationId": "74d719561741d8f0cf67db2cd1664ca1",
                "parameters": [
                    {
                        "name": "x-token",
                        "in": "header",
                        "description": "Key API",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "names or doc",
                        "required": false,
                        "example": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                }
            },
            "post": {
                "tags": [
                    "Person"
                ],
                "summary": "Create Person",
                "operationId": "e3fa321f1f16f618321269761b9dce45",
                "parameters": [
                    {
                        "name": "x-token",
                        "in": "header",
                        "description": "Key API",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Request Body Description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "examples": {
                                "result": {
                                    "summary": "An result object.",
                                    "value": {
                                        "full_name": "",
                                        "surnames": "",
                                        "type_doc": "",
                                        "number_doc": ""
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                }
            }
        },
...
        "/api/possibility_failure": {
            "get": {
                "tags": [
                    "PossibilityFailure"
                ],
                "summary": "Get all Status",
                "operationId": "9105ac3224cc22b627001966e668eecd",
                "parameters": [
                    {
                        "name": "x-token",
                        "in": "header",
                        "description": "Key API",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get account user information"
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                }
            }
        },
        "/api/legal-process": {
            "get": {
                "tags": [
                    "Legal-Process"
                ],
                "summary": "Get all process or Search",
                "operationId": "4b207a71dfe974855c3507a7b1079a1d",
                "parameters": [
                    {
                        "name": "x-token",
                        "in": "header",
                        "description": "Key API",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "",
                        "required": false,
                        "example": ""
                    }
                ],
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                }
            },
            "post": {
                "tags": [
                    "Legal-Process"
                ],
                "summary": "Create Process",
                "operationId": "31ccf9cf523e07f6b2e35566dea28c33",
                "parameters": [
                    {
                        "name": "x-token",
                        "in": "header",
                        "description": "Key API",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Request Body Description",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "examples": {
                                "result": {
                                    "summary": "An result object.",
                                    "value": {
                                        "city": "123456",
                                        "office_id": 1,
                                        "demanding_id": 1,
                                        "defendant_id": 1,
                                        "attorney_id": 2,
                                        "niu": "123",
                                        "reference_internal": "123",
                                        "reference_external": "123",
                                        "facts": "123",
                                        "class_procces_id": 1,
                                        "action_id": 1,
                                        "status_id": 1,
                                        "failure_possibility_id": 1,
                                        "failure_possibility": 1
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": ""
                    },
                    "default": {
                        "description": "An error has occurred."
                    }
                }
            }
        },
...
    },
...
}